Refactor: type 클래스 패키징
