+++ src/main/java/com/zero/triptalk/planner/type/SortType.java
@@ -1,0 +1,7 @@
+package com.zero.triptalk.planner.type;
+
+public enum SortType {
+    RECENT,
+    LIKES,
+    VIEWS
+}
